Refactor: login edit
