+++ feature/src/main/java/com/example/feature/ui/onboarding/AgreementScreen.kt
@@ -1,8 +1,9 @@
-package com.example.feature.onboarding
+package com.example.feature.ui.onboarding
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
@@ -12,13 +13,11 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
 import com.example.core.ui.theme.*
-import com.example.core.ui.theme.Pretendard
 import com.example.feature.R
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 
 @Composable
@@ -33,6 +32,9 @@ fun AgreementScreen(navController: NavHostController) {
         if (allAgree) {
             agreeTerms = true
             agreePrivacy = true
+        } else {
+            agreeTerms = false
+            agreePrivacy = false
         }
     }
 
@@ -75,21 +77,18 @@ fun AgreementScreen(navController: NavHostController) {
             color = HelloWorldGrayScale300
         )
 
-        Spacer(modifier = Modifier.height(32.dp))
+        Spacer(modifier = Modifier.height(32.dp),)
 
-        Column(modifier = Modifier.fillMaxWidth()) {
+        Column(modifier = Modifier.fillMaxWidth(),
+            verticalArrangement = Arrangement.spacedBy(12.dp)) {
             AgreementCheckbox(text = "모두 동의", checked = allAgree, onCheckedChange = {
                 allAgree = it
             })
-
-            Spacer(modifier = Modifier.height(12.dp))
 
             AgreementCheckbox(text = "개인정보 처리방침", checked = agreePrivacy, onCheckedChange = {
                 agreePrivacy = it
                 allAgree = agreePrivacy && agreeTerms
             })
-
-            Spacer(modifier = Modifier.height(12.dp))
 
             AgreementCheckbox(text = "서비스 이용약관", checked = agreeTerms, onCheckedChange = {
                 agreeTerms = it
@@ -185,7 +184,7 @@ fun TermsModal(onDismiss: () -> Unit) {
                         onClick = {
                             agreed.value = true
                             coroutineScope.launch {
-                                kotlinx.coroutines.delay(100)
+                                delay(100)
                                 onDismiss()
                             }
                         },
